pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = ["yurii <y-kostyuk03@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.32.1"
requests = "^2.32.3"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
loguru = "^0.7.3"
passlib = "^1.7.4"
alembic = "^1.14.0"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.10.3"}
python-multipart = "^0.0.19"
bcrypt = "^4.2.1"
pytest = "^8.3.4"
httpx = "0.27.2"
pytest-asyncio = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"

